# api/api.yaml
openapi: 3.0.3
info:
  title: VCBackend
  description: Backend component to verify credentials
  version: 0.0.1
...
paths:
...
  /verifier/api/v1/receivecredential/{state}:
    get:
      tags:
        - verifier
        - frontend
      parameters:
        - $ref: '#/components/parameters/State'
      operationId: VerifierPageReceiveCredential
      summary: Presents a rendered page, showing the credential associated with the requested state
      responses:
        '200':
          description: A rendered page, showing the credential
          content:
            text/html:
              schema:
                type: object
...
  /verifier/api/v1/poll/{state}:
    get:
      tags:
        - verifier
        - api
      parameters:
        - $ref: '#/components/parameters/State'
      operationId: VerifierAPIPoll
      summary: Endpoint to check the current state of a session
      responses:
        '200':
          description: Returns either the state of the session(pending, expired) or the credential itself
          content:
            text/plain:
              schema:
                type: string
  /verifier/api/v1/token/{state}:
    get:
      tags:
        - verifier
        - api
      parameters:
        - $ref: '#/components/parameters/State'
      operationId: VerifierAPIToken
      summary: Return the jwt for the requested session
      responses:
        '200':
          description: The access token
          content:
            text/plain:
              schema: 
                type: string
        '403':
          description: Unauthorized, in case the session is not authorized
  /verifier/api/v1/startsiop:
    get:
      tags:
        - verifier
        - api 
      parameters:
        - $ref: '#/components/parameters/QueryState'
      operationId: VerifierAPIStartSIOP
      summary: Initiates the siop flow and returns the 'openid://...' connection string
      responses:
        '200': 
          description: The connection string 
          content:
            text/plain:
              schema:
                type: string
                example: openid://?scope=dsba.credentials.presentation.PacketDeliveryService&response_type=vp_token&response_mode=post&client_id=did:key:z6MktZy7CErCqdLvknH6g9YNVpWupuBNBNovsBrj4DFGn4R1&redirect_uri=http://localhost:3000/verifier/api/v1/authenticationresponse&state=&nonce=BfEte4DFdlmdO7a_fBiXTw
...
  /verifier/api/v1/authenticationresponse:
    post:
      tags:
        - verifier
        - api 
      parameters:
        - $ref: '#/components/parameters/QueryState'
      operationId: VerifierAPIAuthenticationResponse
      summary: Stores the credential for the given session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VCContainer'
      responses:
        '200':  
          description: Ok when it worked
          content:
            text/plain:
              schema:
                type: string
                enum: 
                  - "ok"
...
  /wallet/api/v1/sendcredential:
    get:
      tags:
        - wallet
        - frontend
      parameters:
        - $ref: '#/components/parameters/QueryState'
        - $ref: '#/components/parameters/QueryId'
        - $ref: '#/components/parameters/RedirectUri'
      operationId: WalletPageSendCredential
      summary: Executes a post with the given credential to the redirect_uri and returns a rendered page, presenting the post-result
      responses:
        '200':
          description: The rendered page
          content:
            text/html:
              schema:
                type: object
...
components:
  parameters:
...
    State:
      name: state
      in: path
      required: true
      schema: 
        type: string
        example: 274e7465-cc9d-4cad-b75f-190db927e56a
    QueryState:
      name: state
      in: query
      required: true
      schema: 
        type: string
        example: 274e7465-cc9d-4cad-b75f-190db927e56a
    QueryId:
      name: id
      in: query
      required: true
      schema:
        type: string
        example: urn:uuid:274e7465-cc9d-4cad-b75f-190db927e56a
    RedirectUri:
      name: redirect_uri
      in: query
      required: true
      schema:
        type: string
        example: https://my-verifier.com/verifier/api/v1/authenticationresponse
  schemas:
...
    CredentialsType:
      type: array
      items:
        type: string
      example:
        - PacketDeliveryService
        - VerifiableCredential
    Did:
      type: object
      properties:
        id:
          type: string
          format: did
          example: did:key:z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh
    Date:
      type: string
      format: date-time
      example: 2023-01-06T07:51:36Z
    CredentialSchema:
      type: object
      properties:
        id: 
          type: string
          example: https://raw.githubusercontent.com/hesusruiz/dsbamvf/main/schemas/PacketDeliveryService/2022-10/schema.json
        type:
          type: string
          example: FullJsonSchemaValidator2021
    Proof:
      type: object
      properties:
        type:
          type: string
          example: JsonWebSignature2020
        creator: 
          type: string
          format: did
          example: did:key:z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh
        created:
          $ref: '#/components/schemas/Date'
        verificationMethod:
          type: string
          example: did:key:z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh#z6Mks9m9ifLwy3JWqH4c57EbBQVS2SpRCjfa79wHb5vWM6vh
        jws:
          type: string
          example: eyJiNjQiOmZhbHNlLCJjcml0IjpbImI2NCJdLCJhbGciOiJFZERTQSJ9..6xSqoZja0NwjF0af9Zknqx3Cbh9GENunBf9C8uL2ulGfwus3UFM_ZnhPjWtHPl-72E9p3BT5f2ptZoYktMKpDA
    SubjectRole:
      type: object
      properties:
        names:
          type: array
          items: 
            type: string
          example:
            - GOLD_CUSTOMER
            - STANDARD_CUSTOMER
        target:
          type: string
          example: did:key:mytarget
    PacketDeliverySubject:
      type: object
      properties:
        id:
          type: string
          format: did
          example: did:key:z6Mkfdio1n9SKoZUtKdr9GTCZsRPbwHN8f7rbJghJRGdCt88
        familyName:
          type: string
          example: Mustermann
        firstName:
          type: string
          example: Max
        roles:
          type: array
          items:
            $ref: '#/components/schemas/SubjectRole'
        email:
          type: string
          format: email
          example: max@mustermann.org
    VerifiableCredential:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/CredentialsType'
        '@context':
          type: array
          items:
            type: string
            format: uri
          example:
            - https://www.w3.org/2018/credentials/v1
            - https://w3id.org/security/suites/jws-2020/v1
        id:
          type: string
          example: urn:uuid:274e7465-cc9d-4cad-b75f-190db927e56a
        issuer:
          $ref: '#/components/schemas/Did'
        issuanceDate:
          $ref: '#/components/schemas/Date'
        issued:
          $ref: '#/components/schemas/Date'
        validFrom:
          $ref: '#/components/schemas/Date'
        credentialSchema:
          $ref: '#/components/schemas/CredentialSchema'
        credentialsSubject:
          $ref: '#/components/schemas/PacketDeliverySubject'
        proof:
          $ref: '#/components/schemas/Proof'
    VCContainer:
      type: object
      properties:
        credential:
          $ref: '#/components/schemas/VerifiableCredential'
...